removed take_human_input.py and switched to bwi_msgs service. closes #20
